Lift voice chat logic to parent component and use toolbarRight instead
